chore(l1_provider): move l1 provider errors to a new l1 provider types crate
