+++ crates/starknet_l1_provider_types/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "starknet_l1_provider_types"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[dependencies]
+async-trait.workspace = true
+chrono = { workspace = true, features = ["serde"] }
+derive_more.workspace = true
+mockall.workspace = true
+papyrus_proc_macros.workspace = true
+serde = { workspace = true, features = ["derive"] }
+starknet_api.workspace = true
+starknet_sequencer_infra.workspace = true
+thiserror.workspace = true
+tracing.workspace = true
+
+
+[lints]
+workspace = true
