+++ crates/starknet_l1_provider_types/src/errors.rs
@@ -1,12 +1,11 @@
-use papyrus_base_layer::ethereum_base_layer_contract::EthereumBaseLayerError;
+use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
-use crate::ProviderState;
+use crate::l1_provider_types::ProviderState;
 
-#[derive(Error, Debug)]
+#[derive(Clone, Debug, Deserialize, Error, Serialize)]
+// TODO(Mohammad): Add `EthereumBaseLayerError` error, and solve the serialization issue.
 pub enum L1ProviderError {
-    #[error(transparent)]
-    BaseLayer(#[from] EthereumBaseLayerError),
     #[error(
         "`get_txs` called while in `Pending` state, likely due to a crash; restart block proposal"
     )]
